python/voc-report/requirements.txt
google-auth-oauthlib==1.2.0
google-api-python-client==2.154.0
pandas==2.2.3
openai==1.41.0
python-dotenv==1.0.1
tabulate==0.9.0
streamlit==1.38.0
